implementing jwt

+++ src/main/java/com/qualitychemicals/qciss/profile/dto/SummaryDto.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class SummaryDto {
     private int id;
-    private int totalSavings;
-    private int totalShares;
-    private int pendingFee;
+    private double totalSavings;
+    private double totalShares;
+    private double pendingFee;
 }
